Under construction: Removing users from season management (community).
