+++ src/app/seasonmanager/update/update-season-model.ts
@@ -1,16 +1,25 @@
 import { Betoffice } from 'src/app/betoffice-json/model/betoffoce-data-model';
-import { CheckableParty } from '../../admin/community/update/checkable-party';
 
 export class UpdateSeasonModel {
 
-  readonly seasonTypes: Array<string> = ['CL', 'CUP', 'EC', 'LEAGUE', 'UEFACUP', 'WC'];
-  readonly teamTypes: Array<string> = ['DFB', 'FIFA'];
+    readonly seasonTypes: Array<string> = ['CL', 'CUP', 'EC', 'LEAGUE', 'UEFACUP', 'WC'];
+    readonly teamTypes: Array<string> = ['DFB', 'FIFA'];
 
-  submitted = false;
-  season: Rest.SeasonJson;
+    submitted = false;
+    season: Rest.SeasonJson = {
+        id: 0,
+        name: '',
+        year: '',
+        openligaLeagueSeason: '',
+        openligaLeagueShortcut: '',
+        currentRoundId: 0,
+        teamType: 'DFB',
+        seasonType: 'LEAGUE',
+        rounds: [],
+    };
 
-  constructor() {
-    this.season = new Betoffice.SeasonModel();
-  }
+    constructor() {
+        this.season = new Betoffice.SeasonModel();
+    }
 
 }
